Allow installing with Latte 3
This relaxes the version bounds so that people can continue to use
the renderers after upgrading to Latte 3.

I did not really bother fixing the latte macros themselves as
they do not really seem to be maintained, only supporting Bootstrap 3.
The existing `latte.macros` config key will just be a no-op with Latte 3,
so Latte will just switch to the default macro definitions from Nette Forms.