Fix Input Macros
The demo has not been updated to use the new Latte extension so the `input` and `label` macros were picked from `Nette\Bridges\FormsLatte\FormsExtension`. But even if we add our extension the macros from `FormsExtension` will take precedence. Let’s add a prefix.

The template nodes themselves passed incorrect arguments – `.var` is not a thing in Latte 3.

On top of that, an incorrect class name would be used `Nextras\FormsRendering\LatteTags` instead of the Bootstrap 3 specific subclass.